requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk.asset-awscli-v1==2.2.200
    # via aws-cdk-lib
aws-cdk.asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk.asset-node-proxy-agent-v5==2.0.165
    # via aws-cdk-lib
aws-cdk-lib==2.87.0
    # via
    #   -r requirements.in
    #   cdk-nag
cattrs==23.1.2
    # via jsii
cdk-nag==2.27.74
    # via -r requirements.in
constructs==10.2.69
    # via
    #   -r requirements.in
    #   aws-cdk-lib
    #   cdk-nag
exceptiongroup==1.1.2
    # via cattrs
importlib-resources==6.0.0
    # via jsii
jsii==1.85.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.7.1
    # via
    #   cattrs
    #   jsii
